refactor: simplify loading recipe steps
To improve clarity, readability, and maintainability, remove the JOIN
for recipes and steps. Since the recipes data structure will have more
member lists, the JOIN complicates this prematurely. It's better to be
clear and readable at this time than have optimal performance. Premature
optimization is the root of all evil as Donald Knuth says.

Instead of a JOIN, query the recipes table for a list of recipes then
for each recipe, query by recipe id for the steps. This obviates the
need for manually converting joined rows into steps.